refactor: split the client code
